% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_table.R
\name{annotate_table}
\alias{annotate_table}
\title{Annotate an initialised table table of 20}
\usage{
annotate_table(fig, script, colors, frametext_size = 16)
}
\arguments{
\item{fig}{Object created by \code{initialise_table()}}

\item{script}{Subset of \code{scripts}}

\item{colors}{Vector of colors}

\item{frametext_size}{Frame-text font size}
}
\description{
Annotate an initialised table table of 20
}
