% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_lookup.R
\name{beta_lookup}
\alias{beta_lookup}
\title{Lookup regression weight for fitted models}
\usage{
beta_lookup(input, predictor, outcome = "overweight-4y")
}
\arguments{
\item{input}{Elected category}

\item{predictor}{Elected predictor name}

\item{outcome}{Elected model name}
}
\value{
One or more regression weights
}
\description{
This function consults the internal \code{betas} data frame
}
\examples{
beta_lookup("Meisje", "Geslacht", "overweight-4y")
beta_lookup("", "Leeftijd vader", "overweight-4y")
}
