% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{betas}
\alias{betas}
\title{Estimated regression weights for predictive models}
\format{
\code{betas}
A data frame with 4052 rows and 3 columns:
\describe{
\item{outcome}{Model name}
\item{Voorspeller}{Predictor name}
\item{Categorie}{Category for discrete variables. Empty of numerical variables.}
\item{Gewicht}{Estimated regression weight}
\item{Modelterm}{Name of model term in the original model (for checking)}
}
}
\source{
Constructed from models fitted by Mirthe Hendriks (2023) using CBS-linked data.
}
\usage{
betas
}
\description{
Estimated regression weights for predictive models
}
\keyword{datasets}
