% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmiChart.R
\name{bmiChartServer}
\alias{bmiChartServer}
\title{Server element for BMI chart}
\usage{
bmiChartServer(id, bmidata, emptyplot)
}
\arguments{
\item{id}{string, shiny id}

\item{bmidata}{reactive data frame}

\item{emptyplot}{the chart canvas to draw on}
}
\description{
Server element for BMI chart
}
