% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{corop}
\alias{corop}
\title{COROP area of The Netherlands}
\format{
\code{corop}
A data frame with 40 rows and 3 columns:
\describe{
\item{corop}{COROP number}
\item{naam}{COROP name}
\item{gemeenten}{Communities}
}
}
\source{
Taken from a file named \code{corop.txt}.
}
\usage{
corop
}
\description{
COROP area of The Netherlands
}
\keyword{datasets}
