% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_framedata.R
\name{create_framedata}
\alias{create_framedata}
\title{Create frame data}
\usage{
create_framedata(outcome = "overweight-4y", seed = NULL)
}
\arguments{
\item{outcome}{Character}

\item{seed}{Seed value}
}
\description{
Creates a data frame with 800 rows divided over 8 animation frames
}
