% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tab10.R
\name{create_tab10}
\alias{create_tab10}
\title{Create a "table of 10" animation from data}
\usage{
create_tab10(
  pri = NULL,
  name = "overweight-4y-3",
  script = NULL,
  palet = c("mandarin", "redgrey"),
  display = c("medium", "large", "default"),
  centile = TRUE,
  seed = NULL,
  ...
)
}
\arguments{
\item{pri}{Numeric. Personal risk estimate (0 < \code{pri} < 1).}

\item{name}{Name of script within built-in scripts \code{tab10::scripts}.
\code{name} is ignored if the \code{script} argument is specified.}

\item{script}{A tibble with the structure of \code{tab10::scripts}.}

\item{palet}{Name of table10 color palette}

\item{display}{Display size, either \code{"medium"} or \code{"large"})}

\item{centile}{Logical. Tooltips as centiles (default) or as absolute risk}

\item{seed}{Seed value}

\item{\dots}{Arguments passed down to \code{\link[=initialise_table]{initialise_table()}}}
}
\description{
Create a "table of 10" animation from data
}
