% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table10.R
\name{create_table10}
\alias{create_table10}
\title{Create a "table of 10" animation from data}
\usage{
create_table10(
  pri,
  name = "overweight-4y-2",
  palet = "mandarin",
  centile = TRUE,
  seed = NULL,
  width = 700,
  height = 900,
  size = 1100,
  wrap = 75
)
}
\arguments{
\item{pri}{Numeric. Personal risk estimate (0 < \code{pri} < 1).}

\item{name}{Name of script}

\item{palet}{Name of table10 color palette}

\item{centile}{Logical. Tooltips as centiles (default) or as absolute risk}

\item{seed}{Seed value}

\item{width}{With of display (px)}

\item{height}{Height of display (px)}

\item{size}{Circle size}

\item{wrap}{Integer. Number of character per line}
}
\description{
Create a "table of 10" animation from data
}
