% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_cases2.R
\name{draw_cases2}
\alias{draw_cases2}
\title{Draws 100 labeled records from the case and prediction vector}
\usage{
draw_cases2(y, p, ntab = 10000L, seed = NULL)
}
\arguments{
\item{y}{Vector of 0 and 1, indicating a non-case and case, respectively}

\item{p}{Vector of predicted probabilities. Same length as \code{y}}

\item{ntab}{Total n in the table of 10}

\item{seed}{Random generated seed for \code{set.seed()}}
}
\description{
Draws 100 labeled records from the case and prediction vector
}
\examples{
pred <- tab10::predictions
y <- pred[pred$outcome == "preterm-32w", "y"]
p <- pred[pred$outcome == "preterm-32w", "p"]
cc <- tab10:::draw_cases2(y, p, ntab = 10000)
sum(cc$y)
}
\author{
Stef van Buuren, Aug 2023
}
