% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{emptyplot}
\alias{emptyplot}
\title{Pre-calculated empty BMI chart}
\format{
\code{emptyplot}
An object of class \code{c("gg", "ggplot")}
}
\usage{
emptyplot
}
\description{
Pre-calculated empty BMI chart
}
\keyword{datasets}
