% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk2rank.R
\name{expit}
\alias{expit}
\title{Inverse logit transform}
\usage{
expit(alpha)
}
\arguments{
\item{alpha}{Linear predictor from logistic regression}
}
\value{
Probability
}
\description{
Inverse logit transform
}
\examples{
expit(0)
}
