% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk2rank.R
\name{p2rank}
\alias{p2rank}
\title{Convert probability to risk rank}
\usage{
p2rank(p, outcome = "overweight-4y")
}
\arguments{
\item{p}{Probability}

\item{outcome}{Name of risk model}
}
\value{
Rank between 1 and 100
}
\description{
Convert probability to risk rank
}
\examples{
p2rank(c(0, 0.15, 0.2, 0.25, 1), "overweight-4y")
}
