% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_palette.R
\name{pick_palette}
\alias{pick_palette}
\title{Return a vector of five colors}
\usage{
pick_palette(name = c("mandarin", "redshadow", "softred"))
}
\arguments{
\item{name}{The name of the palette}
}
\description{
Return a vector of five colors
}
