% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_bar.R
\name{probability_bar}
\alias{probability_bar}
\title{Create coordinates to visualise the probability of an event}
\usage{
probability_bar(n, xlim = c(0, 1), ylim = c(0, 1), m = 10, info = TRUE)
}
\arguments{
\item{n}{Sample size}

\item{xlim}{Numerical vector of length 2 with minimum and maximum probability}

\item{ylim}{Numerical vector of length 2 regulating high of the bar}

\item{m}{number of points to try each step}

\item{info}{logical. Should minimum distance between printed}
}
\value{
A data frame with columns \code{x}, \code{y} and \code{case} (0 or 1)
}
\description{
Create coordinates to visualise the probability of an event
}
\examples{
p <- probability_bar(n = 1000, ylim = c(0, 0.2))
col <- c("navy", "red")[p$case + 1]
MASS::eqscplot(x = p$x, y = p$y, pch = 19, cex = 0.7, col = col)
}
