% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{risk_rank_data}
\alias{risk_rank_data}
\title{Dataset with processed risk and rank scores}
\format{
A data frame with columns:
\describe{
\item{outcome}{The outcome of the model}
\item{p}{Risk estimate, 0 < p < 1}
\item{pct}{rank}
}
}
\usage{
risk_rank_data
}
\description{
Note 1: 100 random draws per outcome
}
\details{
Note 2: Generated with seed = 1
}
\keyword{datasets}
