% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\docType{data}
\name{scripts}
\alias{scripts}
\title{Table of Ten scripts}
\format{
\subsection{\code{scripts}}{

A tibble with columns:
\describe{
\item{out}{The outcome of the model}
\item{lng}{Language, ISO 639-1 code}
\item{vsn}{Character, version number}
\item{lab}{Label of entry}
\item{txt}{The text of the entry}
}
}
}
\usage{
scripts
}
\description{
Script text used to populate Table of Ten fields.
}
\keyword{datasets}
