% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table10.R
\name{create_table10}
\alias{create_table10}
\title{Create a "table of 10" animation from data}
\usage{
create_table10(
  script = "overweight-4y-2",
  y = NULL,
  p = NULL,
  palet = "mandarin",
  centile = TRUE,
  seed = 123,
  width = 700,
  height = 900,
  size = 1100
)
}
\arguments{
\item{script}{Name of script}

\item{y}{Vector of 0 and 1, indicating a non-case and case, respectively}

\item{p}{Vector of predicted probabilities. Same length as \code{y}}

\item{palet}{Name of table10 color palette}

\item{centile}{Logical. Tooltips as centiles (default) or as absolute risk}

\item{seed}{Seed value}

\item{width}{With of display (px)}

\item{height}{Height of display (px)}

\item{size}{Circle size}
}
\description{
Create a "table of 10" animation from data
}
