% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ncases.R
\name{draw_cases}
\alias{draw_cases}
\title{Draws 100 labeled records from the case and prediction vector}
\usage{
draw_cases(y, p, probs = seq(0, 1, 0.1), ntab = 100L)
}
\arguments{
\item{y}{Vector of 0 and 1, indicating a non-case and case, respectively}

\item{p}{Vector of predicted probabilities. Same length as \code{y}}

\item{probs}{Vector of cut points to define group quantiles}

\item{ntab}{Total n in the table of 10}
}
\description{
Draws 100 labeled records from the case and prediction vector
}
\author{
Stef van Buuren, Aug 2023
}
