% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_cases2.R
\name{draw_cases2}
\alias{draw_cases2}
\title{Draws 100 labeled records from the case and prediction vector}
\usage{
draw_cases2(y, p, ntab = 100L, seed = NULL)
}
\arguments{
\item{y}{Vector of 0 and 1, indicating a non-case and case, respectively}

\item{p}{Vector of predicted probabilities. Same length as \code{y}}

\item{ntab}{Total n in the table of 10}

\item{seed}{Random generated seed for \code{set.seed()}}
}
\description{
Draws 100 labeled records from the case and prediction vector
}
\examples{
set.seed(1)
y <- rbinom(2716, 1, 0.2)
p <- runif(2716, max = 0.5)
cc <- tab10:::draw_cases2(y, p)

}
\author{
Stef van Buuren, Aug 2023
}
