% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise_table.R
\name{initialise_table}
\alias{initialise_table}
\title{Initialise a table of 10}
\usage{
initialise_table(
  data,
  script,
  colors,
  width = 700,
  height = 900,
  size = 1100,
  frametext_size = 16,
  xaxis_tickfont_size = 16,
  xaxis_titlefont_size = 20,
  yshift = -100
)
}
\arguments{
\item{data}{Dataset created by \code{make_coordinates()}}

\item{script}{Subset of \code{scripts}}

\item{colors}{Vector of colors}

\item{width}{With of display (px)}

\item{height}{Height of display (px)}

\item{size}{Circle size}

\item{frametext_size}{Frame-text font size}

\item{xaxis_tickfont_size}{X-axis tick font size}

\item{xaxis_titlefont_size}{X-axis title font size}

\item{yshift}{Frame-text box vertical parameter}
}
\description{
Initialise a table of 10
}
