% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{predictions}
\alias{predictions}
\title{Datasets with model predictions}
\format{
\subsection{\code{scripts}}{

A tibble with columns:
\describe{
\item{outcome}{The outcome of the model}
\item{y}{0 = no overweight 4y, 1 = overweight 4y}
\item{p}{Risk estimate, 0 < p < 1}
}
}
}
\usage{
predictions
}
\description{
Datasets with model predictions
}
\keyword{datasets}
