% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{scripts}
\alias{scripts}
\title{Table of Ten scripts}
\format{
\subsection{\code{scripts}}{

A tibble with columns:
\describe{
\item{name}{Script name}
\item{outcome}{The outcome of the model}
\item{version}{Character, version number}
\item{language}{Language, ISO 639-1 code}
\item{frame}{Integer, frame number}
\item{show}{Logical, show the frame?}
\item{label}{Label of entry}
\item{text}{The text of the entry}
}
}
}
\usage{
scripts
}
\description{
Script text used to populate Table of Ten fields.
}
\keyword{datasets}
